requirements.txt
streamlit
pyyaml
tensorflow
opencv-python
joblib
matplotlib
sklearn